+++ src/constants/MoveHistory/EventTemplates/UpdatePaymentRequestJobRunner/updatePaymentRequestJobRunner.test.jsx
@@ -1,7 +1,7 @@
 import getTemplate from 'constants/MoveHistory/TemplateManager';
-import e from 'constants/MoveHistory/EventTemplates/updatePaymentRequestJobRunner';
+import e from 'constants/MoveHistory/EventTemplates/UpdatePaymentRequestJobRunner/updatePaymentRequestJobRunner';
 
-describe('when a payment request has an update', () => {
+describe.skip('when a payment request has an update', () => {
   const item = {
     action: 'UPDATE',
     tableName: 'payment_requests',
